+++ test/components/tab_nav_component_test.rb
@@ -125,4 +125,20 @@ class PrimerTabNavComponentTest < Minitest::Test
       assert_text("Extra")
     end
   end
+
+  def test_renders_extra_content_after_the_tabs
+    render_inline(Primer::TabNavComponent.new(label: "label")) do |c|
+      c.tab(selected: true) { "Tab 1" }
+      c.tab { "Tab 2" }
+      c.extra(align: :right) { "Extra" }
+    end
+
+    assert_selector(".tabnav") do
+      assert_selector("nav.tabnav-tabs[aria-label='label']") do
+        assert_selector("a.tabnav-tab[aria-current='page']", text: "Tab 1")
+        assert_selector("a.tabnav-tab", text: "Tab 2")
+      end
+      assert_text("Extra")
+    end
+  end
 end